+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Updated `bin/grav clean` command to remove unnecessary vendor files (save some bytes) 
     * Added a `http_status_code` Twig function to allow setting HTTP status codes from Twig directly.
     * Deter XSS attacks via URI path/uri methods (credit:newbthenewbd)
+    * Added support for `$uri->toArray()` and `(string)$uri`
 
 # v1.3.4
 ## 09/29/2017